### 2017-03-30 #107 @AliciaSchep
iheatmapr
Makes complex, interactive heatmaps. The package includes a modular system for iteratively building up complex heatmaps, as well as the `iheatmap` function for making relatively standard heatmaps.

```
Package: iheatmapr
Title: Interactive Complex Heatmaps
Version: 0.2.4
Authors@R: c(person("Alicia", "Schep", email = "aschep@gmail.com",
                  role = c("aut", "cre")),
              person("Sarah","Kummerfeld", email = "kummerfeld.sarah@gene.com",
                  role = c("aut")),
              person("Genentech", role = "cph"))
Description: iheatmapr is a package for making complex, interactive heatmaps.
    The package includes a modular system for iteratively building up complex
    heatmaps, as well as the iheatmap function for making relatively standard
    heatmaps.
    plotly (>= 4.1.1),
    plyr,
    S4Vectors,
    ggdendro,
    fastcluster,
    RColorBrewer,
    htmlwidgets,
    scales,
    knitr
    htmltools,
    datasets,
    shiny,
    reshape2,
    roxygen2,
    covr
biocViews: Visualization
Collate:
    'AllClasses.R'
    'AllGenerics.R'
    'annotations.R'
    'as_plotly.R'
    'axes.R'
    'axis_labels.R'
    'axis_titles.R'
    'barplot.R'
    'clustering.R'
    'clusters.R'
    'colorbars.R'
    'colors.R'
    'components.R'
    'dendogram.R'
    'generic_subplots.R'
    'groups.R'
    'main_heatmap.R'
    'iheatmap.R'
    'iheatmapr.R'
    'layout.R'
    'shiny_click_test.R'
    'signal.R'
    'subplot.R'
    'summary.R'
    'utils.R'
```

https://github.com/AliciaSchep/iheatmapr

Anyone who wants to visualize data using heatmaps.  Package is not intended to be domain specific, although some fields tend to use heatmaps more than others.

There are great tools in R for creating relatively simple interactive heatmaps  (plotly, d3heatmap, heatmaply) or creating static complex heatmaps (ComplexHeatmap). However, there are no tools (that I am aware of) facilitating easy creation of complex, interactive heatmaps.
