### 2017-03-30 #107 @maelle
### Editor checks:

---

#### Editor comments

Thanks for your submission @AliciaSchep! :grinning:  Currently looking for reviewers.

A few comments:

* The repository doesn't have a DOI yet but I guess you'd archive it after the review?

* Here is the output from `goodpractice::gp()` that can help you and the reviewers

```r
It is good practice to

  ✖ write unit tests for all functions, and all package code in
    general. 78% of code lines are covered by test cases.

    R/AllGenerics.R:93:NA
    R/AllGenerics.R:96:NA
    R/AllGenerics.R:141:NA
    R/AllGenerics.R:144:NA
    R/AllGenerics.R:201:NA
    ... and 587 more lines

  ✖ not use "Depends" in DESCRIPTION, as it can cause name clashes,
    and poor interaction with other packages. Use "Imports" instead.
  ✖ avoid long code lines, it is bad for readability. Also, many
    people prefer editor windows that are about 80 characters wide.
    Try make your lines shorter than 80 characters

    R\AllGenerics.R:34:1
    R\axes.R:389:1
    R\axes.R:390:1
    R\axes.R:510:1
    R\axes.R:630:1
    ... and 37 more lines

  ✖ not import packages as a whole, as this can cause name clashes
    between the imported packages. Instead, import only the specific
    functions you need.
  ✖ fix this R CMD check NOTE: Namespace in Imports field not
    imported from: 'fastcluster' All declared Imports should be
    used.
  ✖ fix this R CMD check NOTE: add_col_clustering,Iheatmap: no
    visible global function definition for 'hclust'
    add_row_clustering,Iheatmap: no visible global function
    definition for 'hclust' Undefined global functions or variables:
    hclust Consider adding importFrom("stats", "hclust") to your
    NAMESPACE file.
```

* Using `devtools::spell_check()` I found the following typos
    * pallette               add_main_heatmap.Rd:37
    * heirarchical           add_col_clustering.Rd:17, add_row_clustering.Rd:17
    * annotion               add_col_annotation.Rd:26, add_row_annotation.Rd:26
---

Reviewers:
Due date:
