+++ NAMESPACE
@@ -58,6 +58,7 @@ import(S4Vectors)
 import(methods)
 importFrom(RColorBrewer,brewer.pal)
 importFrom(RColorBrewer,brewer.pal.info)
+importFrom(fastcluster,hclust)
 importFrom(htmlwidgets,createWidget)
 importFrom(htmlwidgets,saveWidget)
 importFrom(htmlwidgets,sizingPolicy)
