### 2017-03-30 024555b
shortened long lines
