+++ tests/testthat/test_dendro.R
@@ -6,7 +6,8 @@ test_that("can add a row signal to single horizontal heatmap",{
 })
 
 test_that("can add a row signal to single vertical heatmap",{
-  test_plot <- main_heatmap(a, orientation = "vertical") %>% add_row_dendro(row_dendro)
+  test_plot <- main_heatmap(a, orientation = "vertical") %>% 
+    add_row_dendro(row_dendro)
   expect_iheatmap(test_plot, "row_dendro_vertical", "vertical")
 })
 
@@ -16,7 +17,8 @@ test_that("can add a column signal to single horizontal heatmap",{
 })
 
 test_that("can add a column signal to single vertical heatmap",{
-  test_plot <- main_heatmap(a, orientation = "vertical") %>% add_col_dendro(col_dendro)
+  test_plot <- main_heatmap(a, orientation = "vertical") %>% 
+    add_col_dendro(col_dendro)
   expect_iheatmap(test_plot, "col_dendro_vertical", "vertical")
 })
 

[…]
