### 2017-03-30 7cfbea9
fixed typos, updated site
