### 2017-03-30 #107 @AliciaSchep
Thanks @maelle!

I have updated the typos, long lines, and some of the NAMESPACE issues.

 For the use of DEPENDS, I think it is appropriate in this case to force user import of plotly. For importing the packages as whole, I am only importing  S4Vectors and methods -- as I am using a lot of functions & classes from these packages I thought that might be appropriate, but am of course open to more guidance on this question.

I will work on adding more tests to get closer to 100% test coverage.
