### 2017-03-30 #107 @maelle
Great but note that 100% coverage isn't compulsory 😉
