### 2017-04-01 #107 @maelle
@AliciaSchep the reviewers are now assigned! 😀

Thanks for agreeing to review this package @andeek @carlganz! 😃

As a reminder here are links to the recently updated [reviewing](https://github.com/ropensci/onboarding/blob/master/reviewing_guide.md) and [packaging](https://github.com/ropensci/onboarding/blob/master/packaging_guide.md) guides and to the [review template](https://raw.githubusercontent.com/ropensci/onboarding/master/reviewer_template.md).
