### 2017-04-04 AliciaSchep/iheatmapr#1 @carlganz
Small typo in documentation
Hello,

I believe in your Roxygen documentation you are missing a `@examples` here: https://github.com/AliciaSchep/iheatmapr/blob/master/R/subplot.R#L39

The example for `add_subplot` isn't appearing as a result.

Kind Regards
