### 2017-04-04 10c109f
added missing examples tag go add_supblot documentation
