+++ R/subplot.R
@@ -37,6 +37,7 @@ setMethod("make_trace", signature = c(x = "GenericPlot"),
 #' @aliases add_subplot,Iheatmap-method
 #' @author Alicia Schep
 #' 
+#' @examples
 #' mat <- matrix(rnorm(24, ncol = 6))
 #' hm <- iheatmap(mat) %>% add_subplot(x = 1:5, y=1:5, side = "top")
 #' 
