+++ man/add_subplot.Rd
@@ -40,15 +40,16 @@ an interactive graphic
 \description{
 Adds an arbitrary subplot to iheatmap
 }
+\examples{
+mat <- matrix(rnorm(24, ncol = 6))
+hm <- iheatmap(mat) \%>\% add_subplot(x = 1:5, y=1:5, side = "top")
+
+# Print heatmap if interactive session 
+if (interactive()) hm 
+}
 \seealso{
 \code{\link{iheatmap}}
 }
 \author{
 Alicia Schep
-
-mat <- matrix(rnorm(24, ncol = 6))
-hm <- iheatmap(mat) %>% add_subplot(x = 1:5, y=1:5, side = "top")
-
-# Print heatmap if interactive session 
-if (interactive()) hm
 }
