### 2017-04-04 AliciaSchep/iheatmapr#1 @AliciaSchep
Thanks!  Have updated the roxygen documentation & recreated the rd file.
