### 2017-04-04 58a43d1
fix typo in add_subplot example
