+++ R/subplot.R
@@ -38,7 +38,7 @@ setMethod("make_trace", signature = c(x = "GenericPlot"),
 #' @author Alicia Schep
 #' 
 #' @examples
-#' mat <- matrix(rnorm(24, ncol = 6))
+#' mat <- matrix(rnorm(24), ncol = 6)
 #' hm <- iheatmap(mat) %>% add_subplot(x = 1:5, y=1:5, side = "top")
 #' 
 #' # Print heatmap if interactive session 
