### 2017-04-06 #107 @carlganz
*Consider adding a more concrete statement of need in the readme. The paper.md and the vignette both provide good statements of need.*

  *Consider mentioning the dependency on `BiocInstaller` since there may be users who are not familiar with how to install packages from `BioConductor`.*

  *Excellent vignette*

  *`setup_colorbar_grid` and `save_iheatmap` do not include examples.*

>##### Paper (for packages co-submitting to JOSS)
>
>The package contains a `paper.md` with:
>
>- [X] **A short summary** describing the high-level functionality of the software
>- [X] **Authors:**  A list of authors with their affiliations
>- [X] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

This package provides R users with a collection of easy-to-use functions for iteratively building complex interactive heatmaps. The `iheatmap` function initiates a heatmap, and a variety of function of the form `add_*` allow thee user to include annotations, barplots, clustering, grouping, dendrograms, labels, plots, titles, and summaries to both the rows, and columns as well as additional heatmaps. The package makes nice use of `magrittr` pipes, and `plotly` graphics, which gives it a nice modern feel.

The package is extremely well documented with an extensive vignette, and examples for almost every function. The package is also well tested with 79% code coverage. It contains a library of graphs, which are recreated, and then compared to in the tests.

Overall the package is very well done, and meets all the rOpenSci packaging guidelines. Functions have proper names, and documentation. Use of `plotly` in depends makes sense.

A few suggestions:

* Add clearer statement of need in readme. The first paragraph nicely explains what `iheatmapr` does, but I don't think it is a clear statement of need. The JOSS paper, and the vignette both make clearer statements of need.

* Consider mentioning how to install `BiocInstaller` in the installation instructions. I'll confess, I had to google how to install `BioConductor` packages because I hadn't done it before.

* Add examples for `setup_colorbar_grid` and `save_iheatmap` (you already fixed AliciaSchep/iheatmapr#1)

* Add DOIs for references in paper.md if they exist

* Consider adding Travis tests for Mac, and App-Veyor tests for Windows.

* Consider adding a badge in readme for minimum R version (i.e. `badgecreatr::minimal_r_version_badge("2.10")`)

* The vignette is very well done. Only suggestion is to maybe explain the differences between `clustering` vs `clusters` vs `groups` since they all seem like similar concepts to me.

Hope that helps.

Carl
