### 2017-04-07 #107 @AliciaSchep
@carlganz thank you for the review and helpful suggestions! I will follow up on all the suggestions and post an update to this thread when that is complete.
