### 2017-04-07 12faa94
updated README to include bioconductor install tip [ci skip]
