### 2017-04-20 #107 @andeek
#### Paper (for packages co-submitting to JOSS)

The package contains a `paper.md` with:

- [X] **A statement of need** clearly staing problems the software is designed to solve and its target audience.

---

Overall, this package is really great. I had a really fun time playing with and and definitely see the importance of the work. I think the modular structure fits really nicely with how R is trending lately (in terms of piping) and I appreciate that that must not have been super straightforward to create. Overall, I think is a wonderful contribution and is very fun as well!

In addition, the vignette was extremely well put together and I appreciate the thorough documentation.

I have a few suggestions and explanations for any tasks above that did not receive an "X".

#### Explanations for no "X"s above and suggestions for improvement

Perhaps you could explain the difference between `add_col_clustering()` and `add_col_dendrogram()`, it looks like they are very similar in the result.

##### Documentation

- [ ] **Examples:** Here is are the functions that I found without examples: `save_iheatmap`, `setup_colorbar_grid`
- [ ] **Community guidelines:** There is a CONDUCT file, but not a CONTRIBUTING file. I'm not sure if you need both.

##### Paper (for packages co-submitting to JOSS)

- [ ] **References:** You are missing some DOI's.

##### Functionality

- [ ] **Performance:** I didn't see any "performance claims", but the package seemed to work well for me.
- [ ] **Automated tests:** I received the following warnings using the `test_that` script:

    ```
    1: In .removePreviousCoerce(class1, class2, where, prevIs) :
      methods currently exist for coercing from "IheatmapMainX" to "IheatmapAxis"; they will be replaced.
    2: In .removePreviousCoerce(class1, class2, where, prevIs) :
      methods currently exist for coercing from "IheatmapMainY" to "IheatmapAxis"; they will be replaced.
    ```

- [ ] **Packaging guidelines**: Avoid starting the description with the package name.
