### 2017-04-21 #107 @AliciaSchep
Thanks @andeek for the review! I hope to have time soon to address all issues in both reviews.
