### 2017-04-30 4e66993
update regex in pname_check to add escapes when needed
