### 2017-04-30 49c0d14
added contributing.md
