### 2017-04-30 583afd1
added r version badge
