+++ paper/paper.bib
@@ -13,7 +13,8 @@
   number={18},
   pages={2847--2849},
   year={2016},
-  publisher={Oxford Univ Press}
+  publisher={Oxford Univ Press},
+  doi={https://doi.org/10.1093/bioinformatics/btw313}
 }
 
 @Manual{d3heatmap,
