### 2017-04-30 004e382
added contributing to rbuildignore
