### 2017-04-30 627d53e
add mac testing to travis
