+++ appveyor.yml
@@ -2,6 +2,7 @@
 
 # Download script file from GitHub
 init:
+  - git config --global core.autocrlf true
   ps: |
         $ErrorActionPreference = "Stop"
         Invoke-WebRequest http://raw.github.com/krlmlr/r-appveyor/master/scripts/appveyor-tool.ps1 -OutFile "..\appveyor-tool.ps1"
