### 2017-05-01 41d481b
update to appveyor.yml to maybe help with expect_equal_to_reference
