### 2017-05-16 #107 @maelle
@AliciaSchep I see you've done some work on the package, do you think you'll soon be able to answer to the reviews? :wink:
