### 2017-05-17 #107 @AliciaSchep
Sorry, I'm behind on everything!  I have made most of the changes suggested, although I still need to figure out why the tests on Appveyor are failing...
