+++ appveyor.yml
@@ -14,6 +14,8 @@ install:
 cache:
   - C:\RLibrary
 
+matrix:
+  - R_ARCH: x64
 # Adapt as necessary starting from here
 
 build_script:
