### 2017-05-17 #107 @maelle
No problem! I work on Windows, so if you need me to test anything, just ask 😉
