+++ tests/testthat/test_adjust_x_domains.R
@@ -1,8 +1,6 @@
 #library(iheatmapr)
 context("adjust_x_domains")
 
-a <- matrix(rnorm(200),ncol=10)
-
 test_that("adding a main heatmap of same size to right gives correct sizing",{
   p <- main_heatmap(a) %>% add_main_heatmap(a, size = 1, buffer = 0.015)
   x1_start <- domain_start(xaxes(p)[["x"]])
