### 2017-05-20 f8bfad6
updated docs and vignette
