+++ R/colorbars.R
@@ -51,6 +51,15 @@ setMethod("zmax", c(x = "DiscreteColorbar"),
 #' @param y_start top most position of colorbar grid
 #' @return \code{\link{IheatmapColorbarGrid-class}} object
 #' @export
+#' @examples 
+#' 
+#' cb_grid <- setup_colorbar_grid(nrows = 2, x_spacing = 0.2)
+#' mat <- matrix(rnorm(24), nrow = 6)
+#' hm <- iheatmap(mat, colorbar_grid = cb_grid, cluster_rows = "kmeans",
+#'          cluster_cols = "kmeans", row_k = 3, col_k = 2)
+#' 
+#' # Print heatmap if interactive session 
+#' if (interactive()) hm 
 setup_colorbar_grid <- function(nrows = 3,
                                 y_length = y_spacing * 0.9,
                                 x_spacing = 0.16,
