### 2017-05-20 #107 @AliciaSchep
I have now addressed almost all of the reviewer concerns, thanks again to @andeek and @carlganz for taking the time to review the package.

- [x] DOIs for references

Most of the references are other packages, and I couldn't find DOI, but I have added the DOI for the reference that is a paper.

- [x] Examples

I added examples for `save_iheatmap` and `setup_colorbar_grid`

- [x] Statement of need in README

I added statement of need into README, similar to what was in vignette and paper.

- [x] Add travis mac testing

- [x] Add appveyor testing for windows [now passes!]

- [x] Add BiocInstaller instructions

Added to README

- [x] Add badge for minimal R

- [x] Clarify clusters vs clustering vs groups functions [Reviewer 1] and clustering vs dendrogram [Reviewer 2]

Edited vignette to make more clear [basic idea is that add_*_clustering functions do the work of clustering for you, but if you want to do clustering yourself there are are lower level functions for adding dendrogram or cluster annotation]

- [x] Add contributing guidelines

- [x] Avoid starting description with package name

Have changed description field to avoid this

**The main comments I haven't resolved:**

- [ ] warning with tests

I haven't been able to reproduce these warnings, so not sure how best to go about addressing them!

- [ ] Any performance claims of the software been confirmed.

README & vignette don't make performance claims, the goal is not better performance than other similar packages, but more features & flexibility.  Should I add performance tests?
