### 2017-05-20 #107 @maelle
Thanks @AliciaSchep! Nice Work!

Regarding what you say about performance do you cite these other packages in a Vignette?  That'd be enough. 😊

@andeek and @carlganz are you happy with the changes? Please tell me if you don't have time to have a look, in which case I'll do that.
