### 2017-05-20 #107 @AliciaSchep
Thanks @maelle. With regards to other related packages,  I cite them in README and paper.md, but not currently in vignette -- I can add that there as well.
