### 2017-05-22 #107 @maelle
@AliciaSchep yep maybe that'd be better because when you install the package from CRAN you often don't see the README&paper. In any case IMO just citing them is enough as regards performance. :-)
