### 2017-05-23 e4c4bdd
update vignette to add reference to other pkg
