### 2017-05-23 #107 @AliciaSchep
Great point about CRAN, hadn't thought of that -- I added mention of related packages to vignette as well.
