### 2017-05-30 d0ce4b4
Fix plotly dependencies
