### 2017-06-01 #107 @andeek
I would like to, but I'm turning in my dissertation to my committee next week. I'm sorry to drop the ball, I understand if you can't wait for that.
