### 2017-06-01 #107 @AliciaSchep
@andeek just turned my dissertation over to my committee last week, so totally understand... I think delay is fine, although will defer to what @maelle thinks is best.  Good luck wrapping up the dissertation!

On a related note, @carlganz and @maelle, the package is currently broken as a result of update in plotly... fix in progress https://github.com/AliciaSchep/iheatmapr/pull/7
