### 2017-06-01 #107 @maelle
Wow congratulations all around for the dissertations @andeek @AliciaSchep! :rocket: :tada:

If Alicia is ok with waiting and if Andee is motivated to have a look at the package again after turning in her dissertation and a break I guess, it's fine. But Andee you can also choose to drop the ball, as you wish.

:four_leaf_clover: for the fix Alicia!
