### 2017-06-04 3e09ad6
remove plotly r package dependency
