### 2017-06-04 65c72e5
updated shiny test
