+++ NAMESPACE
@@ -15,6 +15,7 @@ export(make_shapes)
 export(make_trace)
 export(renderIheatmap)
 export(setup_colorbar_grid)
+export(test_iheatmapr_event)
 exportClasses(Iheatmap)
 exportClasses(IheatmapAnnotation)
 exportClasses(IheatmapAxis)
