### 2017-06-04 df79242
Update CONTRIBUTING.md
fix typo
