### 2017-06-04 c185910
Update CONTRIBUTING.md
fixed more typos
