### 2017-06-04 544e7ec
Update NEWS.md
