+++ NEWS.md
@@ -1,3 +1,7 @@
+# iheatmapr 0.3.0
+
+* Removed plotly r package dependency.  iheatmapr now directly interfaces with plotlyjs instead. While this may mean having two copies of the plotly.js library, the benefit is that the htmlwidget for iheatmapr can be customized for plots created by iheatmapr and that iheatmapr now has more control over when the js component gets updated.
+
 # iheatmapr 0.2.1
 
 * Added a `NEWS.md` file to track changes to the package.
