### 2017-06-08 #107 @AliciaSchep
@maelle - yes!  I ended up making quite a few changes so as to remove the plotly R package dependency so iheatmapr can have more control over version of plotly.js and how to pass data to plotly.js, and less reliance on internal functioning of plotly R package, which can change. This has additional benefit of allowing for iheatmapr specific shiny handlers -- I added a section to vignette about shiny integration.
