### 2017-06-08 #107 @carlganz
I only had a few small issues, and they have all been addressed:

* Added two examples

* Improved statement of need

* Expanded installation instructions

* References don't have DOIs

I also took a look at the additional javascript code associated with the HTMLwidget, and it is very well done.

I've checked everything off in my review above. Let me know if you need any additional feedback.
