### 2017-06-10 #107 @andeek
Hi all,

I've had a second pass and everything looks great. All of my questions were addressed.  I also had a short look at the JS code and it looks very nice.

I do have one question: By changing to using the JS library (instead of the R package), should you reference the JS library instead of the R package in the paper?

Other than that, everything looks very nice. Easy to install now, and all tests pass on my machine. Great job! Let me know if I need to do anything more.
