### 2017-06-10 #107 @AliciaSchep
Thanks @carlganz and @andeek!

Re the new JS code, the main reason it's nice is that it was derived from a previous version of the plotly R package JS code -- I put a note at the top `// Function adapted from Plotly R Package 3.60`, but perhaps I should give additional credit elsewhere?  I was thinking of adding an acknowledgements section to README and/or vignette, could mention it there or more formally in package metadata.

Also, since I am now including the JS code from plotly should I adjust the package level author, copyright, and license info to reflect that? The plotly LICENSE is included in the inst/htmlwidgets/lib/plotlyjs folder.  For another package that I submitted to Bioconductor that included a C++ library from a third party, I included an [AUTHORS](https://github.com/GreenleafLab/motifmatchr/blob/master/inst/AUTHORS) file and [COPYRIGHTS](https://github.com/GreenleafLab/motifmatchr/blob/master/inst/COPYRIGHTS) file to very clear indicate that the C++ library was the work of other authors.  Would a similar strategy be appropriate here as well?

Re updating the paper to reference JS library, I agree that it would be appropriate to cite the JS library but think that it should still also cite the R package.  I'll update to cite both.
