### 2017-06-15 #107 @maelle
@AliciaSchep Looking at the package right now, looks great. Could you add the rOpenSci footer to the bottom of the README? This is the code `[![ropensci_footer](http://ropensci.org/public_images/github_footer.png)](http://ropensci.org)` .

Reg. your question about the new JS code, was it settled or should we try finding an answer via the Slack or forum? I think that's the last thing before accepting your package.

After acceptance I'll ask you to transfer the repo to rOpenSci, and you'll need to:
* Update the URLs.
* Do a new release with a new DOI for then proceeding with the JOSS submission.

But first let's solve this JS code issue. :wink:
