### 2017-06-16 #107 @AliciaSchep
Thanks @maelle.  I have not yet figured out the JS code, been meaning to post a question, but haven't gotten around to it yet.  Any suggestion of best place to post question?  Not very familiar with the Slack, so don't know what 'channel' is appropriate
