+++ README.md
@@ -62,4 +62,4 @@ Hovering over the heatmaps yields a tooltip with the name of the row and column 
 
 # Documentation
 
-See the [vignette](https://aliciaschep.github.io/iheatmapr/articles/iheatmapr.html) for a more thorough introduction to the package.
+See the [vignette](https://aliciaschep.github.io/iheatmapr/articles/full_vignettes/iheatmapr.html) for a more thorough introduction to the package.
