### 2017-07-17 a6351bf
Added plotly authors, with comment that plotly.js, and acknowledgments
