+++ DESCRIPTION
@@ -2,11 +2,36 @@ Package: iheatmapr
 Type: Package
 Title: Interactive, Complex Heatmaps
 Version: 0.3.0
+Author: Alicia Schep, Sarah Kummerfeld 
 Authors@R: c(person("Alicia", "Schep", email = "aschep@gmail.com",
                   role = c("aut", "cre")),
               person("Sarah","Kummerfeld", email = "kummerfeld.sarah@gene.com",
                   role = c("aut")),
-              person("Genentech", role = "cph"))
+              person("Genentech", role = "cph"),
+              person("Alex", "Johnson", role = "ctb",
+                  comment = "plotly.js library"),
+              person("Étienne ", "Tétreault-Pinard", role = "ctb",
+                  comment = "plotly.js library"),
+              person("Mikola", "Lysenko", role = "ctb",
+                  comment = "plotly.js library"),  
+              person("Ricky", "Reusser", role = "ctb",
+                  comment = "plotly.js library"),   
+              person("Robert", "Monfera", role = "ctb",
+                  comment = "plotly.js library"),    
+              person("Nicolas", "Riesco", role = "ctb",
+                  comment = "plotly.js library"),  
+              person("Miklós", "Tusz", role = "ctb",
+                  comment = "plotly.js library"),    
+              person("Chelsea", "Douglas", role = "ctb",
+                  comment = "plotly.js library"), 
+              person("Ben", "Postlethwaite", role = "ctb",
+                  comment = "plotly.js library"),   
+              person("Chris", "Parmer", role = "ctb",
+                  comment = "plotly.js library"),    
+              person("Alex", "Vados", role = "ctb",
+                  comment = "plotly.js library"), 
+              person("Plotly", role = "cph",
+                  comment = "plotly.js library"))
 Description: Make complex, interactive heatmaps. iheatmapr includes a modular 
     system for iteratively building up complex heatmaps, as well as the iheatmap 
     function for making relatively standard heatmaps.
@@ -43,6 +68,7 @@ RoxygenNote: 6.0.1
 VignetteBuilder: knitr
 URL: https://github.com/AliciaSchep/iheatmapr
 BugReports: https://github.com/AliciaSchep/iheatmapr/issues
+Encoding: UTF-8
 Collate:
     'AllClasses.R'
     'AllGenerics.R'
