### 2017-07-17 caadc14
add summary function options
