+++ R/summary.R
@@ -17,6 +17,8 @@
 #' @param xname internal name of xaxis
 #' @param yname internal name of yaxis
 #' @param type scatter or bar?
+#' @param summary_function summary function to use, default is mean, options 
+#' are mean, median, sd, var, mad, max, min
 #' @param ... additional arguments to \code{\link{add_row_plot}} or 
 #' \code{\link{add_row_barplot}}
 #' 
@@ -67,10 +69,12 @@ setMethod(add_row_summary, c(p = "Iheatmap"),
                    xname = NULL,
                    yname = current_yaxis(p),
                    type = c("scatter","bar"),
+                   summary_function = c("mean","median","sd","var","mad","max","min"),
                    ...){
             
             side <- match.arg(side)
             type <- match.arg(type)
+            summary_function <- match.arg(summary_function)
             
             if (is.null(heatmap_name)){
               hm <- get_heatmap(p, xname = xname, yname = yname, side = side)
@@ -93,7 +97,7 @@ setMethod(add_row_summary, c(p = "Iheatmap"),
             mat <- get_data(hm)
             
             row_summary_without_groups <- function(p){
-              x <- rowMeans(mat, na.rm = TRUE)
+              x <- apply(mat, 1, summary_function, na.rm = TRUE)
               p <- if (type == "scatter"){
                 add_row_plot(p,
                                 x = x, 
@@ -125,7 +129,7 @@ setMethod(add_row_summary, c(p = "Iheatmap"),
             row_summary_with_groups <- function(p, groups, colors){
               for (i in seq_along(levels(groups))){
                 sel <- which(groups == levels(groups)[i])
-                x <- rowMeans(mat[,sel,drop = FALSE],na.rm=TRUE)
+                x <- apply(mat[,sel,drop = FALSE], 1, summary_function, na.rm = TRUE)
                 p <- if (type == "scatter"){
                   add_row_plot(p,
                                   x = x, color = colors[i], 
@@ -214,6 +218,8 @@ setMethod(add_row_summary, c(p = "Iheatmap"),
 #' @param xname internal name of xaxis
 #' @param yname internal name of yaxis
 #' @param type scatter or bar?
+#' @param summary_function summary function to use, default is mean, options 
+#' are mean, median, sd, var, mad, max, min
 #' @param ... additional arguments to \code{\link{add_col_plot}} or 
 #' \code{\link{add_col_barplot}}
 #' 
@@ -264,10 +270,12 @@ setMethod(add_col_summary, c(p = "Iheatmap"),
                    xname = current_xaxis(p),
                    yname = NULL,
                    type = c("scatter","bar"),
+                   summary_function = c("mean","median","sd","var","mad","max","min"),
                    ...){
             
             side <- match.arg(side)
             type <- match.arg(type)
+            summary_function <- match.arg(summary_function)
             
             if (is.null(heatmap_name)){
               hm <- get_heatmap(p, xname = xname, yname = yname, side = side)
@@ -290,7 +298,7 @@ setMethod(add_col_summary, c(p = "Iheatmap"),
             mat <- get_data(hm)
             
             col_summary_without_groups <- function(p){
-              y <- colMeans(mat, na.rm = TRUE)
+              y <- apply(mat, 2, summary_function, na.rm = TRUE)
               
               p <- if (type == "scatter"){
                 add_col_plot(p,
@@ -323,7 +331,7 @@ setMethod(add_col_summary, c(p = "Iheatmap"),
             col_summary_with_groups <- function(p, groups, colors){
               for (i in seq_along(levels(groups))){
                 sel <- which(groups == levels(groups)[i])
-                y <- colMeans(mat[sel,,drop = FALSE],na.rm=TRUE)
[…]
