+++ man/add_row_summary.Rd
@@ -9,7 +9,9 @@
 \S4method{add_row_summary}{Iheatmap}(p, groups = NULL, heatmap_name = NULL,
   colors = NULL, tracename = "Row Summary", showlegend = FALSE,
   side = c("right", "left"), layout = list(), size = 0.3, buffer = 0.02,
-  xname = NULL, yname = current_yaxis(p), type = c("scatter", "bar"), ...)
+  xname = NULL, yname = current_yaxis(p), type = c("scatter", "bar"),
+  summary_function = c("mean", "median", "sd", "var", "mad", "max", "min"),
+  ...)
 }
 \arguments{
 \item{p}{\code{\link{Iheatmap-class}} object}
@@ -40,6 +42,9 @@ size of first heatmap}
 \item{yname}{internal name of yaxis}
 
 \item{type}{scatter or bar?}
+
+\item{summary_function}{summary function to use, default is mean, options 
+are mean, median, sd, var, mad, max, min}
 
 \item{...}{additional arguments to \code{\link{add_row_plot}} or 
 \code{\link{add_row_barplot}}}
