+++ man/add_col_summary.Rd
@@ -9,7 +9,9 @@
 \S4method{add_col_summary}{Iheatmap}(p, groups = NULL, heatmap_name = NULL,
   colors = NULL, tracename = "Col Summary", showlegend = FALSE,
   side = c("top", "bottom"), layout = list(), size = 0.3, buffer = 0.02,
-  xname = current_xaxis(p), yname = NULL, type = c("scatter", "bar"), ...)
+  xname = current_xaxis(p), yname = NULL, type = c("scatter", "bar"),
+  summary_function = c("mean", "median", "sd", "var", "mad", "max", "min"),
+  ...)
 }
 \arguments{
 \item{p}{\code{\link{Iheatmap-class}} object}
@@ -39,6 +41,9 @@ size of first heatmap}
 \item{yname}{internal name of yaxis}
 
 \item{type}{scatter or bar?}
+
+\item{summary_function}{summary function to use, default is mean, options 
+are mean, median, sd, var, mad, max, min}
 
 \item{...}{additional arguments to \code{\link{add_col_plot}} or 
 \code{\link{add_col_barplot}}}
