+++ vignettes/full_vignettes/iheatmapr.Rmd
@@ -381,6 +381,16 @@ main_heatmap(Indometh_matrix) %>%
   add_col_summary(groups = TRUE) 
 ```
 
+By default, summarization is done using `mean`. However, the `summary_function` argument can be set to 'median', 'sd', 'var', 'mad', 'max', or 'min' in order to perform any of those alternate summarizations. We can also use the `layout` argument to pass a title to the new axis.    
+
+```{r}
+main_heatmap(Indometh_matrix) %>% 
+  add_row_clustering(k = 3) %>% 
+  add_col_summary(summary_function = "sd", 
+                  layout = list(title = "sd")) 
+```
+
+
 ## Add barplot
 
 The functions `add_col_barplot` and `add_row_barplot` add barplots along the columns or rows of a main heatmap.  
