### 2017-07-17 401e1a0
added ropensci footer
