### 2017-07-17 #107 @AliciaSchep
@maelle, I finally updated the DESCRIPTION to add the plotly.js authors, using comment to indicate that they are authors of the js package, as suggested by Jim on ropensci slack and similar to what was done in package you suggested.  I also added an acknowledgment section, in which I included a thanks to plotly r package for code that was adapted for this package.  Let me know if you have additional suggestions for making sure appropriate credit is being given!  I also added the ropensci footer
