+++ README.md
@@ -2,6 +2,7 @@
 [![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/AliciaSchep/iheatmapr?branch=master&svg=true)](https://ci.appveyor.com/project/AliciaSchep/iheatmapr)
 [![codecov](https://codecov.io/gh/AliciaSchep/iheatmapr/branch/master/graph/badge.svg?token=CTupoUlXNI)](https://codecov.io/gh/AliciaSchep/iheatmapr)
 ![R version](https://img.shields.io/badge/R%20%3E%3D-3.2.0-blue.svg)
+[![](https://ropensci.org/badges/107_status.svg)](https://github.com/ropensci/onboarding/issues/107)
 
 # iheatmapr
 
