### 2017-07-17 #107 @AliciaSchep
woohoo!  I transferred the repo to ropensci and added the peer review badge.  @maelle  For CI, what are the new links?  Anything else I need to do setup the CI?  I noticed that OpenCPU CI sent me an email saying build had failed, due to BiocInstaller not being installed (package has one BioC dependency, S4Vectors).
