### 2017-07-17 #107 @AliciaSchep
@carlganz and @andeek may I add you as "rev" to the package description?  thanks again for reviewing the package! (feel free to let me know either here or ropensci slack or email-- aschep is my gmail address)
