### 2017-07-17 e6ee46a
update travis badge on readme
