### 2017-07-18 #107 @maelle
@AliciaSchep awesome, thanks for transferring it!

* For OpenCPU let me pink @jeroen

* For Appveyor here is the new badge `[![Build status](https://ci.appveyor.com/api/projects/status/ds0vrh047h196vqj?svg=true)](https://ci.appveyor.com/project/ropensci/iheatmapr)`

* For Travis you just need to replace your username by "ropensci" in the old badge.
