### 2017-07-18 #107 @jeroen
Don't worry about OpenCPU, it doesn't really support BioConductor yet :)
