### 2017-07-18 #107 @maelle
Thanks @jeroen

@AliciaSchep I forgot to mention to update the links in DESCRIPTION
