### 2017-07-18 #107 @AliciaSchep
I updated the links in Description & README.  I tried doing the zenodo thing, but the repository doesn't show up in the list of repos to be able to archive?  Is there some kind of access that has to be set @maelle?
