### 2017-07-19 #107 @AliciaSchep
Thanks @maelle, it shows up now!  Question about timing of things -- I want to now submit package to CRAN, should I do that before doing the DOI & JOSS submission?  Does it matter?
