### 2017-07-20 #107 @maelle
Hi @AliciaSchep, as far as I know it doesn't matter but if you wait before the CRAN submission then you can have the JOSS paper in citation so that users of the CRAN version know how to cite your package. 😊
