### 2017-07-25 c0a1748
Updated ropensci review badge URL
