### 2017-08-03 #107 @maelle
@AliciaSchep Are you going to submit the package to JOSS soon?

For info yesterday I submitted a package to CRAN after JOSS acceptance, this is how CITATION looks like. https://github.com/ropensci/rtimicropem/blob/master/inst/CITATION

If you do the optional item of asking reviewers whether they'd like to be included in DESCRIPTION as "rev", when submitting the package to CRAN you'll have to build it on R-devel (or wait until R3.5 :-)) because otherwise this non-standard role isn't translated well from Authors@R to Authors.
