### 2017-08-04 5573a8e
Add layout= argument to main_heatmap() function
I had to put the font= argument inside a layout= argument to get the code to work for me without error.
