### 2017-08-04 7a15920
export generic for to_widget
