+++ NAMESPACE
@@ -16,6 +16,7 @@ export(make_trace)
 export(renderIheatmap)
 export(setup_colorbar_grid)
 export(test_iheatmapr_event)
+export(to_widget)
 exportClasses(Iheatmap)
 exportClasses(IheatmapAnnotation)
 exportClasses(IheatmapAxis)
