### 2017-08-10 #107 @maelle
@AliciaSchep any plan to submit the package to JOSS soon?
