### 2017-08-10 cbb24e3
Add DOI to readme
