+++ DESCRIPTION
@@ -8,6 +8,10 @@ Authors@R: c(person("Alicia", "Schep", email = "aschep@gmail.com",
               person("Sarah","Kummerfeld", email = "kummerfeld.sarah@gene.com",
                   role = c("aut")),
               person("Genentech", role = "cph"),
+              person("Andee","Kaplan", role = "rev",
+                  comment = "Reviewer for rOpensci review process https://github.com/ropensci/onboarding/issues/107"),
+              person("Carl","Ganz", role = "rev",
+                  comment = "Reviewer for rOpensci review process https://github.com/ropensci/onboarding/issues/107"),    
               person("Alex", "Johnson", role = "ctb",
                   comment = "plotly.js library"),
               person("Étienne ", "Tétreault-Pinard", role = "ctb",
