### 2017-08-10 #107 @AliciaSchep
I have added the reviewers to the DESCRIPTION and made a release.  For JOSS submission, there is a dropdown of suggested editors -- any recommendation there?
