### 2017-08-10 8e7b78d
bumped version number and rebuilt doc website
