### 2017-08-11 #107 @maelle
I actually don't think it makes a difference but when I submitted a paper to JOSS I chose Karthik because he's an rOpenSci editor as well :wink:
