### 2017-08-11 #107 @maelle
Did you ask/inform the reviewers that they were added to DESCRIPTION? I've never seen anyone say no but better to ask. :-)
