### 2017-08-11 #107 @AliciaSchep
Yes I asked the reviewers, both were fine with it!
