### 2017-08-11 #107 @AliciaSchep
And now have also submitted to JOSS!  Thanks @maelle
